Foi implementado ApiError e uma mensagem de erro ao tentar deletar qual quer usuario que não existe do sistema;

+++ api/src/main/java/com/company/api/service/FornecedorService.java
@@ -1,10 +1,14 @@
 package com.company.api.service;
 
+import com.company.api.config.ApiErrorEnum;
 import com.company.api.domain.Fornecedor;
 import com.company.api.repository.FornecedorRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FornecedorService {
@@ -26,6 +30,11 @@ public class FornecedorService {
     public Fornecedor uptade(Integer id, Fornecedor fornecedor) { fornecedor.setId(id); return fornecedorRepository.save(fornecedor); }
 
     public void delete(Integer id) {
+        Optional<Fornecedor> possivelFornecedor = fornecedorRepository.findById(id);
+        if(!possivelFornecedor.isPresent()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    ApiErrorEnum.RECORD_NOT_FOUND_MESSAGE.getDescricao());
+        }
         fornecedorRepository.deleteById(id);
     }
